pyproject.toml
[project]
name = "pr-review-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "khatsukoves",email = "eldarslavikovich@gmail.com"}
]
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi (>=0.128.7,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pygithub (>=2.8.1,<3.0.0)",
    "llama-index (>=0.14.13,<0.15.0)",
    "llama-index-core (>=0.14.13,<0.15.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
